Change the roi_data form

+++ include/rm_detector/detector.h
@@ -63,7 +63,7 @@ public:
   float bbox_conf_thresh_;
   std_msgs::Float32MultiArray roi_data_;
   std::vector<cv::Point2f> roi_point_vec_;
-  cv::Point2i roi_data_point_;
+  cv::Point2f roi_data_point_;
   cv::Mat_<float> discoeffs_;
   cv::Mat_<float> camera_matrix_;
   std::vector<float> discoeffs_vec_;
